Refactor loading bar into own component
