+++ src/components/LoadingBar/index.tsx
